# be-users/build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SupportedColors": {
				"type": "string",
				"enum": [
					"red",
					"blue",
					"green"
				]
			},
			"ITag": {
				"properties": {
					"label": {
						"type": "string"
					},
					"color": {
						"$ref": "#/components/schemas/SupportedColors",
						"description": "supported web colors"
					}
				},
				"required": [
					"label",
					"color"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPost": {
				"properties": {
					"sver": {
						"type": "string",
						"description": "used to keep track of current entity version"
					},
					"title": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"tags": {
						"items": {
							"$ref": "#/components/schemas/ITag"
						},
						"type": "array"
					},
					"created_by": {
						"type": "string",
						"description": "user id reference"
					}
				},
				"required": [
					"title",
					"content",
					"tags"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"description": "User objects allow you to associate actions performed in the system with the user that performed them.\nThe User object contains common information across every user in the system regardless of status and role.",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string",
						"description": "The email the user used to register his account"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"Happy",
							"Sad"
						]
					},
					"phoneNumbers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"posts": {
						"items": {
							"$ref": "#/components/schemas/IPost"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"phoneNumbers"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"id": "52907745-7672-470e-a803-a2f8feb52944",
					"name": "John Doe",
					"phoneNumbers": []
				}
			},
			"Pick_User.email-or-name-or-phoneNumbers_": {
				"properties": {
					"email": {
						"type": "string",
						"description": "The email the user used to register his account"
					},
					"name": {
						"type": "string"
					},
					"phoneNumbers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"email",
					"name",
					"phoneNumbers"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserCreationParams": {
				"$ref": "#/components/schemas/Pick_User.email-or-name-or-phoneNumbers_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "be-users",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}